+++ src/js/popup.js
@@ -54,3 +54,11 @@ function localizeHtmlPage() {
 }
 
 localizeHtmlPage();
+
+document.querySelector('#go-to-options').addEventListener('click', function () {
+  if (chrome.runtime.openOptionsPage) {
+    chrome.runtime.openOptionsPage();
+  } else {
+    window.open(chrome.runtime.getURL('../html/options.html'));
+  }
+});